[FEATURE] single모드 api 설계 전체 완료

See merge request s11-webmobile1-sub2/S11P12E202!89

+++ backend/single/src/main/java/com/scf/single/domain/entity/ContentCheckUser.java
@@ -27,11 +27,10 @@ public class ContentCheckUser {
     private int complete;
 
     @Column(name = "member_id")
-    private int memberId;
+    private Long memberId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "content_id", nullable = false)
     private Content content;
-
 
 }
